update helper methods
